requirements.txt
box2d-py==2.3.8
cloudpickle==1.2.2
cycler==0.10.0
future==0.18.0
gym==0.15.3
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
pyglet==1.3.2
pyparsing==2.4.2
python-dateutil==2.8.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
sklearn==0.0
torch==1.3.0
